ci: add job to check increase in binary size
This job checks the difference in binary size across commits in a PR,
and fails if the difference is more than a set maximum limit. This
allows us to detect early on if there is any possible bloating in the
binary due to poorly structured dependencies. The job can be overridden
by adding a `bloat-ok` label to the PR.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>