refactor: move table rendering into own funcs
The logic for rendering tables with the checkpoint data was all present
in a single function `showContainerCheckpoint()`. This has now been
refactored to provide each table its own function that can be called to
render it. This increases code maintainability going forward as more and
more tables are added to the `checkpointctl show` command.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>