ci: use Ubuntu for tests instead of Fedora
Previously, the Fedora container image was being used to test
checkpointctl since the Ubuntu image had an outdated version of bats.
This does not affect the tests themselves, as the coverage workflow
continues to use `ubuntu-latest` without issues. This change will
eliminate the need for having a privileged container to run criu while
generating image files for tests.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>