Simplified 'list' command code
The code for the 'list' command introduced code which already existed in
the repository. This commit re-uses existing code.

Also, this extends the 'list' command to get the information from
checkpoints created by containerd.

This also changes how the 'list' command works. Before it was always
using '/var/lib/kubelet/checkpoints/' to look for checkpoints and '-p'
extended the search path. Now 'checkpoint list' looks in the default
location and 'checkpoint list /tmp /etc' only looks in '/tmp/' and
'/etc', but no longer in the default location. Always looking in the
default location broke local tests for me.

Signed-off-by: Adrian Reber <areber@redhat.com>