Additional container metadata from containerd
containerd structures the internal metadata differently. For a restore
the status.dump file is also necessary.

The status.dump file contains the content of:

  crictl inspect <CTR-ID> | jq ".status"

and the information is needed during restore to get container
annotations and metadata. CRI-O in contrast has this data available
as metadata in config.dump. As containerd has this information somewhere
else it needs to be saved during checkpointing for a successful restore.

Signed-off-by: Adrian Reber <areber@redhat.com>