fix(ci): move from Fedora to Ubuntu
Previously, we were using Fedora as Ubuntu did not contain the updated
binaries for bats. CRIU was also outdated on the main repositories.
Due to a recent change in the way GitHub runners handle privileged
containers, the old Fedora setup no longer works. This is the easiest
solution that allows us to retain sudo-based local tests that can also
be reused on CI.

Signed-off-by: Prajwal S N <prajwalnadig21@gmail.com>